+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased:
 
+* [**Phil Dibowitz**:](https://github.com/jaymzh)
+  Use lsblk instead of blkid if available.
 
 ## Last Release: 7.2.0